pyproject.toml
[project]
name = "blender-data-sampling"
version = "0.1.0"
description = "Synthetic object detection dataset generation from Blender scenes"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "joblib>=1.3,<2",
  "numpy>=1.23,<2",
  "opencv-python>=4.7,<5",
  "pydantic>=2.6,<3",
  "PyYAML>=6,<7",
  "scikit-learn>=1.3,<2",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"
